fix(#1197): address reviewer findings — delete dead middleware, fix race, fix break
- Delete internal/middleware/health.go: HealthHandler and HealthMiddleware
  have zero production callers since the new vibewarden_health Caddy module
  superseded them. Split ReadyHandler/ReadyMiddleware/ReadyResponse into
  internal/middleware/ready.go first to preserve the live readiness handler.
  Stale "healthy"/"unhealthy" vocabulary in a public file is a tripwire for
  future devs/agents grepping for health-endpoint examples.

- Delete internal/middleware/health_test.go, health_upstream_test.go,
  health_deps_test.go: all tested the now-deleted HealthHandler.

- Remove TestReadyAndHealthAreIndependent from ready_test.go: the test
  called the deleted HealthHandler; the invariant it verified (liveness vs
  readiness independence) is now covered by the Caddy handler tests.

- Fix data race in health_handler_integration_test.go: var returnOK = true
  was shared between the test goroutine (writer) and the httptest server
  handler goroutine (reader) without synchronisation. Changed to
  atomic.Bool with Store/Load. go test -race now passes cleanly.

- Remove dead ServeHTTP call in serveHealthStr: the helper created a
  recorder, called h.ServeHTTP, discarded the recorder, then called
  serveHealth which made its own independent HTTP call. The first call
  was dead code that could mask probe-state flakiness.

- Fix break placement in wiring_serve.go plugin search loop: the break
  fired on any Collector() match, not only on a successful
  MetricsCollectorWithUpstreamHealth assertion. Moved break inside the
  ok2 block so the loop continues if the assertion fails.
