--- internal/middleware/health.go
@@ -1,301 +1,0 @@
-package middleware
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-	"sync"
-	"time"
-
-	"github.com/vibewarden/vibewarden/internal/ports"
-)
-
-// depCacheTTL is how long dependency health check results are cached.
-// This prevents hammering dependencies on every health endpoint request.
-const depCacheTTL = 5 * time.Second
-
-// depProbeTimeout is the per-dependency probe timeout.
-const depProbeTimeout = 3 * time.Second
-
-// ComponentStatus holds the per-component health strings included in the
-// health response under the "components" key.
-type ComponentStatus struct {
-	// Sidecar is the health of the VibeWarden sidecar itself.
-	// It is always "ok" when the sidecar is able to serve this response.
-	Sidecar string `json:"sidecar"`
-
-	// Upstream is the health of the upstream application as reported by the
-	// background health checker. Possible values: "healthy", "unhealthy",
-	// "unknown". Omitted when no upstream health checker is configured.
-	Upstream string `json:"upstream,omitempty"`
-}
-
-// HealthResponse is the JSON response from the health endpoint.
-type HealthResponse struct {
-	// Status is the overall health status: "ok", "degraded", or "unhealthy".
-	Status string `json:"status"`
-
-	// Version is the running VibeWarden binary version.
-	Version string `json:"version"`
-
-	// Components contains per-component health status.
-	Components ComponentStatus `json:"components"`
-
-	// Dependencies maps dependency name to its current health status.
-	// Only present when the health handler was constructed with checkers.
-	Dependencies map[string]ports.DependencyStatus `json:"dependencies,omitempty"`
-}
-
-// cachedEntry holds a single cached dependency status with a fetch timestamp.
-type cachedEntry struct {
-	status    ports.DependencyStatus
-	fetchedAt time.Time
-}
-
-// depStatusCache caches dependency health check results with a fixed TTL.
-// It is safe for concurrent use.
-type depStatusCache struct {
-	mu      sync.Mutex
-	entries map[string]cachedEntry
-	ttl     time.Duration
-}
-
-func newDepStatusCache(ttl time.Duration) *depStatusCache {
-	return &depStatusCache{
-		entries: make(map[string]cachedEntry),
-		ttl:     ttl,
-	}
-}
-
-// get returns the cached DependencyStatus for name when it is still within
-// the TTL, and false otherwise.
-func (c *depStatusCache) get(name string) (ports.DependencyStatus, bool) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	e, ok := c.entries[name]
-	if !ok || time.Since(e.fetchedAt) > c.ttl {
-		return ports.DependencyStatus{}, false
-	}
-	return e.status, true
-}
-
-// set stores a fresh DependencyStatus for name.
-func (c *depStatusCache) set(name string, status ports.DependencyStatus) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.entries[name] = cachedEntry{status: status, fetchedAt: time.Now()}
-}
-
-// HealthHandler returns an http.HandlerFunc for the health check endpoint.
-// The health endpoint is served at /_vibewarden/health.
-//
-// When checkers are provided each dependency is probed on demand (with a
-// 5-second cache to avoid hammering dependencies on every request) and the
-// results are included in the JSON response.
-//
-// When an upstream health checker is provided its most-recently-observed
-// status is included in the response under components.upstream.
-//
-// The overall "status" field is derived from component and dependency results:
-//   - "ok"       — all components and dependencies healthy (or none configured)
-//   - "degraded" — one or more dependencies or the upstream unhealthy; sidecar
-//     is still serving traffic (HTTP 200)
-//   - "unhealthy" — sidecar itself has critical issues (HTTP 503)
-//
-// HTTP status codes: 200 for "ok" and "degraded", 503 for "unhealthy".
-func HealthHandler(version string, upstreamChecker ports.UpstreamHealthChecker, checkers ...ports.DependencyChecker) http.HandlerFunc {
-	cache := newDepStatusCache(depCacheTTL)
-
-	return func(w http.ResponseWriter, r *http.Request) {
-		resp := HealthResponse{
-			Status:  string(ports.HealthSummaryOK),
-			Version: version,
-			Components: ComponentStatus{
-				Sidecar: "ok",
-			},
-		}
-
-		// Populate upstream component status from the background checker.
-		allComponentsHealthy := true
-		if upstreamChecker != nil {
-			upstreamStatus := upstreamChecker.CurrentStatus().String()
-			resp.Components.Upstream = upstreamStatus
-			if upstreamStatus != "healthy" {
-				allComponentsHealthy = false
-			}
-		}
-
-		if len(checkers) > 0 {
-			deps := make(map[string]ports.DependencyStatus, len(checkers))
-			allDepsHealthy := true
-
-			for _, checker := range checkers {
-				name := checker.DependencyName()
-
-				status, hit := cache.get(name)
-				if !hit {
-					// Probe with a bounded timeout so a slow dependency cannot
-					// hold up the health endpoint indefinitely.
-					probeCtx, cancel := context.WithTimeout(r.Context(), depProbeTimeout)
-					status = checker.CheckDependency(probeCtx)
-					cancel()
-					cache.set(name, status)
-				}
-
-				deps[name] = status
-				if status.Status != "healthy" {
-					allDepsHealthy = false
-				}
-			}
-
-			resp.Dependencies = deps
-			if !allDepsHealthy {
-				allComponentsHealthy = false
-			}
-		}
-
-		if !allComponentsHealthy {
-			resp.Status = string(ports.HealthSummaryDegraded)
-		}
-
-		httpStatus := http.StatusOK
-		if resp.Status == string(ports.HealthSummaryUnhealthy) {
-			httpStatus = http.StatusServiceUnavailable
-		}
-
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(httpStatus)
-
-		if err := json.NewEncoder(w).Encode(resp); err != nil {
-			// Best-effort: headers already sent, cannot write error response.
-			return
-		}
-	}
-}
-
-// HealthMiddleware intercepts requests to /_vibewarden/health and serves
-// the health response. All other requests pass through to the next handler.
-//
-// upstreamChecker may be nil when upstream health monitoring is not configured.
-// When dependency checkers are provided they are forwarded to HealthHandler
-// for live dependency status reporting.
-func HealthMiddleware(version string, upstreamChecker ports.UpstreamHealthChecker, checkers ...ports.DependencyChecker) func(next http.Handler) http.Handler {
-	healthHandler := HealthHandler(version, upstreamChecker, checkers...)
-
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path == "/_vibewarden/health" {
-				healthHandler(w, r)
-				return
-			}
-			next.ServeHTTP(w, r)
-		})
-	}
-}
-
-// ReadyResponse is the JSON body returned by the /_vibewarden/ready endpoint.
-type ReadyResponse struct {
-	// Ready is true when all plugins are initialised and the upstream is reachable.
-	Ready bool `json:"ready"`
-
-	// Plugins maps each plugin name to its current health status string.
-	// Possible values: "healthy", "unhealthy".
-	Plugins map[string]string `json:"plugins,omitempty"`
-
-	// Upstream is the current upstream reachability status.
-	// Possible values: "reachable", "unreachable". Omitted when no upstream
-	// health checker is configured.
-	Upstream string `json:"upstream,omitempty"`
-
-	// Degraded maps the name of each non-critical plugin that failed Init or
-	// Start to the failure reason. A non-empty Degraded map means the sidecar
-	// is running in partial-start mode: critical plugins are healthy, but one
-	// or more observability/feature plugins could not start.
-	Degraded map[string]string `json:"degraded,omitempty"`
-}
-
-// ReadyHandler returns an http.HandlerFunc for the readiness probe endpoint
-// served at /_vibewarden/ready.
-//
-// The endpoint returns HTTP 200 with {"ready":true} only when all plugins are
-// healthy and the upstream application is reachable. It returns HTTP 503 with
-// {"ready":false} in any other case.
-//
-// readinessChecker may be nil; when nil the handler always returns 200 {"ready":true}
-// because no plugins or upstream are being monitored.
-func ReadyHandler(readinessChecker ports.ReadinessChecker) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		resp := ReadyResponse{Ready: true}
-
-		if readinessChecker != nil {
-			rs := readinessChecker.ReadinessStatus()
-			resp.Ready = rs.PluginsReady && rs.UpstreamReachable
-
-			// Populate per-plugin status.
-			if len(rs.Plugins) > 0 {
-				resp.Plugins = make(map[string]string, len(rs.Plugins))
-				for name, hs := range rs.Plugins {
-					if hs.Healthy {
-						resp.Plugins[name] = "healthy"
-					} else {
-						resp.Plugins[name] = "unhealthy"
-					}
-				}
-			}
-
-			// Populate upstream status only when an upstream checker is present.
-			// UpstreamReachable being false with no checker means no checker was
-			// configured, not that the upstream is down. We distinguish by checking
-			// whether any plugins are registered — but the cleaner signal is whether
-			// the ReadinessStatus explicitly carries upstream information.
-			// We rely on the ReadinessStatus.UpstreamReachable field: if the
-			// implementation sets it to true unconditionally when no upstream
-			// checker exists, we still want to omit the field from the JSON. To
-			// support that, ReadyHandler accepts a second optional signal via the
-			// readinessChecker interface itself — but since the interface doesn't
-			// expose "has upstream checker", we include the upstream field whenever
-			// a readinessChecker is provided.
-			if rs.UpstreamReachable {
-				resp.Upstream = "reachable"
-			} else {
-				resp.Upstream = "unreachable"
-			}
-
-			// Populate degraded plugin map when partial-start mode is active.
-			if len(rs.DegradedPlugins) > 0 {
-				resp.Degraded = rs.DegradedPlugins
-			}
-		}
-
-		httpStatus := http.StatusOK
-		if !resp.Ready {
-			httpStatus = http.StatusServiceUnavailable
-		}
-
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(httpStatus)
-
-		if err := json.NewEncoder(w).Encode(resp); err != nil {
-			// Best-effort: headers already sent, cannot write error response.
-			return
-		}
-	}
-}
-
-// ReadyMiddleware intercepts requests to /_vibewarden/ready and serves the
-// readiness response. All other requests pass through to the next handler.
-//
-// readinessChecker may be nil; when nil the endpoint always returns 200.
-func ReadyMiddleware(readinessChecker ports.ReadinessChecker) func(next http.Handler) http.Handler {
-	readyHandler := ReadyHandler(readinessChecker)
-
-	return func(next http.Handler) http.Handler {
-		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			if r.URL.Path == "/_vibewarden/ready" {
-				readyHandler(w, r)
-				return
-			}
-			next.ServeHTTP(w, r)
-		})
-	}
-}
