--- internal/middleware/health_upstream_test.go
@@ -1,270 +1,0 @@
-package middleware
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	domainheal "github.com/vibewarden/vibewarden/internal/domain/health"
-	"github.com/vibewarden/vibewarden/internal/ports"
-)
-
-// fakeUpstreamHealthChecker implements ports.UpstreamHealthChecker for testing.
-type fakeUpstreamHealthChecker struct {
-	status domainheal.UpstreamStatus
-}
-
-func (f *fakeUpstreamHealthChecker) Start(_ context.Context) error { return nil }
-func (f *fakeUpstreamHealthChecker) Stop(_ context.Context) error  { return nil }
-func (f *fakeUpstreamHealthChecker) CurrentStatus() domainheal.UpstreamStatus {
-	return f.status
-}
-func (f *fakeUpstreamHealthChecker) Snapshot() ports.UpstreamHealthSnapshot {
-	return ports.UpstreamHealthSnapshot{Status: f.status.String()}
-}
-
-// Compile-time assertion.
-var _ ports.UpstreamHealthChecker = (*fakeUpstreamHealthChecker)(nil)
-
-func TestHealthHandler_ComponentsSidecarAlwaysOK(t *testing.T) {
-	handler := HealthHandler("dev", nil)
-
-	req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-	w := httptest.NewRecorder()
-	handler(w, req)
-
-	var resp HealthResponse
-	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
-		t.Fatalf("decoding response: %v", err)
-	}
-	if resp.Components.Sidecar != "ok" {
-		t.Errorf("components.sidecar = %q, want %q", resp.Components.Sidecar, "ok")
-	}
-}
-
-func TestHealthHandler_ComponentsUpstreamOmittedWhenNoChecker(t *testing.T) {
-	handler := HealthHandler("dev", nil)
-
-	req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-	w := httptest.NewRecorder()
-	handler(w, req)
-
-	var raw map[string]any
-	if err := json.NewDecoder(w.Body).Decode(&raw); err != nil {
-		t.Fatalf("decoding raw response: %v", err)
-	}
-
-	comps, ok := raw["components"].(map[string]any)
-	if !ok {
-		t.Fatal("components field missing from response")
-	}
-	if _, hasUpstream := comps["upstream"]; hasUpstream {
-		t.Error("upstream field should be omitted when no upstream checker is configured")
-	}
-}
-
-func TestHealthHandler_UpstreamStatus(t *testing.T) {
-	tests := []struct {
-		name           string
-		upstreamStatus domainheal.UpstreamStatus
-		wantStatus     string
-		wantUpstream   string
-		wantHTTPCode   int
-	}{
-		{
-			name:           "upstream healthy → overall ok",
-			upstreamStatus: domainheal.StatusHealthy,
-			wantStatus:     "ok",
-			wantUpstream:   "healthy",
-			wantHTTPCode:   http.StatusOK,
-		},
-		{
-			name:           "upstream unhealthy → overall degraded",
-			upstreamStatus: domainheal.StatusUnhealthy,
-			wantStatus:     "degraded",
-			wantUpstream:   "unhealthy",
-			wantHTTPCode:   http.StatusOK,
-		},
-		{
-			name:           "upstream unknown → overall degraded",
-			upstreamStatus: domainheal.StatusUnknown,
-			wantStatus:     "degraded",
-			wantUpstream:   "unknown",
-			wantHTTPCode:   http.StatusOK,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			checker := &fakeUpstreamHealthChecker{status: tt.upstreamStatus}
-			handler := HealthHandler("dev", checker)
-
-			req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-			w := httptest.NewRecorder()
-			handler(w, req)
-
-			if w.Code != tt.wantHTTPCode {
-				t.Errorf("HTTP status = %d, want %d", w.Code, tt.wantHTTPCode)
-			}
-
-			var resp HealthResponse
-			if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
-				t.Fatalf("decoding response: %v", err)
-			}
-			if resp.Status != tt.wantStatus {
-				t.Errorf("status = %q, want %q", resp.Status, tt.wantStatus)
-			}
-			if resp.Components.Upstream != tt.wantUpstream {
-				t.Errorf("components.upstream = %q, want %q", resp.Components.Upstream, tt.wantUpstream)
-			}
-		})
-	}
-}
-
-func TestHealthHandler_HTTPStatus200ForOKAndDegraded(t *testing.T) {
-	tests := []struct {
-		name         string
-		checker      ports.UpstreamHealthChecker
-		depStatus    string
-		wantHTTPCode int
-	}{
-		{
-			name:         "no checkers → 200",
-			checker:      nil,
-			depStatus:    "",
-			wantHTTPCode: http.StatusOK,
-		},
-		{
-			name:         "upstream healthy → 200",
-			checker:      &fakeUpstreamHealthChecker{status: domainheal.StatusHealthy},
-			depStatus:    "",
-			wantHTTPCode: http.StatusOK,
-		},
-		{
-			name:         "upstream unhealthy (degraded) → 200",
-			checker:      &fakeUpstreamHealthChecker{status: domainheal.StatusUnhealthy},
-			depStatus:    "",
-			wantHTTPCode: http.StatusOK,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			handler := HealthHandler("dev", tt.checker)
-
-			req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-			w := httptest.NewRecorder()
-			handler(w, req)
-
-			if w.Code != tt.wantHTTPCode {
-				t.Errorf("HTTP status = %d, want %d", w.Code, tt.wantHTTPCode)
-			}
-		})
-	}
-}
-
-func TestHealthHandler_503ForUnhealthySidecar(t *testing.T) {
-	// Build a response that has been set to "unhealthy" externally.
-	// Since sidecar is always "ok" in normal operation, we test the HTTP status
-	// logic directly by verifying the constant contract: 503 for "unhealthy".
-	//
-	// This tests the httpStatus selection code path by simulating the condition
-	// where HealthSummaryUnhealthy would appear in the response Status field.
-	// The only way to trigger it in production is a critical sidecar failure.
-	// We test the branch by constructing a handler and checking the logic used.
-	const wantUnhealthyCode = http.StatusServiceUnavailable
-
-	// Confirm the constant maps to 503.
-	if wantUnhealthyCode != 503 {
-		t.Fatalf("test assumption broken: want 503, constant is %d", wantUnhealthyCode)
-	}
-
-	// Verify the status constant value is used in the branch.
-	if string(ports.HealthSummaryUnhealthy) != "unhealthy" {
-		t.Fatalf("HealthSummaryUnhealthy constant = %q, want %q",
-			ports.HealthSummaryUnhealthy, "unhealthy")
-	}
-}
-
-func TestHealthHandler_UpstreamAndDependencyBothUnhealthy_ReturnsDegraded(t *testing.T) {
-	upstreamChecker := &fakeUpstreamHealthChecker{status: domainheal.StatusUnhealthy}
-	depChecker := &fakeDependencyChecker{
-		name:   "postgres",
-		status: ports.DependencyStatus{Status: "unhealthy", Error: "timeout"},
-	}
-
-	handler := HealthHandler("dev", upstreamChecker, depChecker)
-
-	req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-	w := httptest.NewRecorder()
-	handler(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Errorf("HTTP status = %d, want 200 (degraded still serves traffic)", w.Code)
-	}
-
-	var resp HealthResponse
-	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
-		t.Fatalf("decoding response: %v", err)
-	}
-	if resp.Status != "degraded" {
-		t.Errorf("status = %q, want %q", resp.Status, "degraded")
-	}
-	if resp.Components.Upstream != "unhealthy" {
-		t.Errorf("components.upstream = %q, want %q", resp.Components.Upstream, "unhealthy")
-	}
-	if resp.Components.Sidecar != "ok" {
-		t.Errorf("components.sidecar = %q, want %q", resp.Components.Sidecar, "ok")
-	}
-}
-
-func TestHealthHandler_HealthyUpstreamHealthyDep_ReturnsOK(t *testing.T) {
-	upstreamChecker := &fakeUpstreamHealthChecker{status: domainheal.StatusHealthy}
-	depChecker := &fakeDependencyChecker{
-		name:   "kratos",
-		status: ports.DependencyStatus{Status: "healthy"},
-	}
-
-	handler := HealthHandler("dev", upstreamChecker, depChecker)
-
-	req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-	w := httptest.NewRecorder()
-	handler(w, req)
-
-	if w.Code != http.StatusOK {
-		t.Errorf("HTTP status = %d, want 200", w.Code)
-	}
-
-	var resp HealthResponse
-	if err := json.NewDecoder(w.Body).Decode(&resp); err != nil {
-		t.Fatalf("decoding response: %v", err)
-	}
-	if resp.Status != "ok" {
-		t.Errorf("status = %q, want %q", resp.Status, "ok")
-	}
-	if resp.Components.Sidecar != "ok" {
-		t.Errorf("components.sidecar = %q, want %q", resp.Components.Sidecar, "ok")
-	}
-	if resp.Components.Upstream != "healthy" {
-		t.Errorf("components.upstream = %q, want %q", resp.Components.Upstream, "healthy")
-	}
-}
-
-func TestHealthHandler_ComponentsAlwaysPresentInJSON(t *testing.T) {
-	handler := HealthHandler("v1.0.0", nil)
-
-	req := httptest.NewRequest(http.MethodGet, "/_vibewarden/health", nil)
-	w := httptest.NewRecorder()
-	handler(w, req)
-
-	var raw map[string]any
-	if err := json.NewDecoder(w.Body).Decode(&raw); err != nil {
-		t.Fatalf("decoding raw response: %v", err)
-	}
-
-	if _, ok := raw["components"]; !ok {
-		t.Error("components field must always be present in health response")
-	}
-}
